+++ Tests/Tests.csproj
@@ -120,11 +120,10 @@
     <Compile Include="EddnTests.cs" />
     <Compile Include="DataDefinitionTests.cs" />
     <Compile Include="GeneratorTests.cs" />
-    <Compile Include="I18NBundleTest.cs" />
+    <Compile Include="I18NTests.cs" />
     <Compile Include="JournalMonitorTests.cs" />
     <Compile Include="BodyScanTests.cs" />
     <Compile Include="ShipTests.cs" />
-    <Compile Include="I18NTests.cs" />
     <Compile Include="VersioningTests.cs" />
     <Compile Include="StarTests.cs" />
     <Compile Include="DataProviderTests.cs" />
